AIPOOL/requirements.txt
# AIPOOL/requirements.txt
# ─────────────────────────────────────────────────────────────────
# Intentionally minimal — no ML, no heavy dependencies.
# ─────────────────────────────────────────────────────────────────
PyYAML>=6.0.1
requests>=2.32.3
urllib3>=2.2.3
certifi>=2024.12.14
